// SurgeonPortal.Api/ClientApp/src/app/api/models/continuouscertification/reference-letter.model.ts
export interface IReferenceLetterModel {
    userId: number;
    official: string;
    roleName: string;
    altRoleName: string;
    roleId: number;
    altRoleId: number;
    explain: string;
    title: string;
    email: string;
    phone: string;
    hosp: string;
    city: string;
    state: string;
    fullName: string;
}